+++ src/Lentille/GitHub/UserPullRequests.hs
@@ -11,6 +11,7 @@ import Lentille.GitHub.GraphQLFragments (fragmentPRData)
 import Lentille.GitHub.Types
 import Lentille.GitHub.Utils
 import Lentille.GraphQL
+import Monocle.Config qualified as Config
 import Monocle.Prelude hiding (id, state)
 import Monocle.Protob.Change
 
@@ -45,7 +46,7 @@ streamUserPullRequests ::
   GraphEffects es =>
   GraphClient ->
   -- A callback to get Ident ID from an alias
-  (Text -> Maybe Text) ->
+  (Text -> Maybe Config.IdentUG) ->
   UTCTime ->
   Text ->
   LentilleStream es Changes
@@ -59,7 +60,7 @@ transformResponse ::
   -- hostname of the provider
   Text ->
   -- A callback to get Ident ID from an alias
-  (Text -> Maybe Text) ->
+  (Text -> Maybe Config.IdentUG) ->
   -- The response payload
   GetUserPullRequests ->
   (PageInfo, Maybe RateLimit, [Text], [Changes])